fix: Correct serialization of command count in SDynamicLayerUnitData
Updated the serialization method to cast the command count to an integer before outputting, ensuring proper data type handling.

Log: fix bug
Bug: https://pms.uniontech.com/bug-view-327629.html

+++ src/drawshape/sitemdata.h
@@ -637,7 +637,7 @@ struct SDynamicLayerUnitData {
         out << layUnit.baseImg;
         out << layUnit.blocked;
         out << layUnit.layerType;
-        out << layUnit.commands.count();
+        out << static_cast<int>(layUnit.commands.count());
         foreach (auto cmd, layUnit.commands) {
             out << cmd->cmdType();
             cmd->serialization(out);
